Revert "[multi]" + git magic
This reverts commit cd78b1b0dd5426f39dfdab07c5e7695be37ad99a.
